Enhance sandboxing
Like discussed, the sandboxing done in
https://github.com/gucci-on-fleek/extractbb/blob/58ae3bf806a21be2f4358bf51106cbf6877dca02/source/extractbb-scratch.lua#L147-L160
is quite nice (and I also think often used).

Also one nice thing is that it modifies `_ENV` instead of modifying `_G`
which makes it impossible to test the code sandboxed. The problem here
is that the sandboxed code might behave differently outside the sandbox.
Thus it is very nice we now can keep the code sandboxed even for
testing.